chore(number-field): move to packages
