+++ packages/components/fields/number-field/src/index.js
@@ -1,0 +1,3 @@
+export { default } from './number-field';
+
+export { default as version } from './version';
