+++ packages/components/fields/number-field/src/number-field.form.story.js
@@ -6,11 +6,11 @@ import { action } from '@storybook/addon-actions';
 import { withKnobs, boolean, select } from '@storybook/addon-knobs/react';
 import Spacings from '@commercetools-uikit/spacings';
 import { PrimaryButton, SecondaryButton } from '@commercetools-uikit/buttons';
-import Section from '../../../../.storybook/decorators/section';
-import FormikBox from '../../../../.storybook/decorators/formik-box';
-import Readme from './README.md';
+import Section from '../../../../../.storybook/decorators/section';
+import FormikBox from '../../../../../.storybook/decorators/formik-box';
+import Readme from '../README.md';
 import NumberField from './number-field';
-import NumberInput from '../../inputs/number-input';
+import NumberInput from '../../../inputs/number-input';
 
 // Cool stuff to try in this story:
 //  - Click the "Age" label and see how the input is focused automatically
