+++ packages/components/fields/number-field/src/number-field.story.js
@@ -10,9 +10,9 @@ import {
   object,
   number,
 } from '@storybook/addon-knobs/react';
-import Section from '../../../../.storybook/decorators/section';
-import Readme from './README.md';
-import * as icons from '../../icons';
+import Section from '../../../../../.storybook/decorators/section';
+import Readme from '../README.md';
+import * as icons from '../../../icons';
 import NumberField from './number-field';
 
 storiesOf('Components|Fields', module)
