+++ packages/components/fields/number-field/src/number-field.visualroute.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { NumberField } from '@commercetools-frontend/ui-kit';
-import { Suite, Spec } from '../../../../test/percy';
+import { Suite, Spec } from '../../../../../test/percy';
 
 const value = '12.50';
