--- src/components/fields/number-field/index.js
@@ -1,1 +1,0 @@
-export { default } from './number-field';
